# [pt][int64_t gemm][fbgemm] Integrating the int64_t GEMM in FBGEMM into PyTorch Linear op